requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-phonenumber-field==8.3.0
phonenumbers==9.0.18
django-cors-headers==4.9.0

